Added 'lineicons' web font and expanded icon set in unicodesMap.json
In this commit, we have introduced a new web font called 'lineicons' version 4.0, located in the `../fonts` directory. The `@font-face` rule has been used to specify the 'lineicons' font family and the path to the font files. We have also added several CSS classes to style the icons, including size, animation, and transformation controls. Furthermore, a large list of icon names and their corresponding unicode values have been added to the `$icons` map, which is used to generate CSS rules for each icon. A `unicode` function has been defined to convert these unicode values to a format usable in CSS content properties. Additionally, the `unicodesMap.json` file has been updated with a large number of key-value pairs, significantly expanding the set of icons that can be referenced in the application.