`Update API fetch URLs in webauthn.js`
Updated the URLs used in the `fetch` method for API calls in the `newregistration()` and `checkregistration()` functions in the `webauthn.js` file. The URLs have been changed from relative paths to absolute paths, pointing to `auth.bloodygang.com`. This suggests that the API calls are now being made to a different server instead of the local server.